# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Music Service",
...
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "YouTube Music API Service",
      "url": "https://github.com/your-repo"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/watch/": {
      "get": {
        "tags": [
          "watch",
          "watch"
        ],
        "summary": "Get watch playlist",
        "description": "Obtiene la playlist de reproducción (siguientes canciones) basada en un video o playlist. Soporta radio y shuffle.",
        "operationId": "get_watch_playlist_api_v1_watch__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID del video para iniciar",
              "examples": [
                "rMbATaj7Il8"
              ],
              "title": "Video Id"
            },
            "description": "ID del video para iniciar"
          },
          {
            "name": "playlist_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID de la playlist",
              "examples": [
                "PL..."
              ],
              "title": "Playlist Id"
            },
            "description": "ID de la playlist"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Número de canciones",
              "examples": [
                25
              ],
              "default": 25,
              "title": "Limit"
            },
            "description": "Número de canciones"
          },
          {
            "name": "start_index",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Índice inicial para paginación",
              "default": 0,
              "title": "Start Index"
            },
            "description": "Índice inicial para paginación"
          },
          {
            "name": "radio",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Obtener playlist de radio",
              "default": false,
              "title": "Radio"
            },
            "description": "Obtener playlist de radio"
          },
          {
            "name": "shuffle",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Obtener playlist en modo shuffle",
              "default": false,
              "title": "Shuffle"
            },
            "description": "Obtener playlist en modo shuffle"
          },
          {
            "name": "include_stream_urls",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Incluir stream URLs y mejores thumbnails para tracks",
              "default": false,
              "title": "Include Stream Urls"
            },
            "description": "Incluir stream URLs y mejores thumbnails para tracks"
          },
          {
            "name": "prefetch_count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": -1,
              "description": "Número de URLs a obtener en paralelo (0 = none, -1 = todas)",
              "default": 10,
              "title": "Prefetch Count"
            },
            "description": "Número de URLs a obtener en paralelo (0 = none, -1 = todas)"
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Watch Playlist Api V1 Watch  Get"
                },
                "example": {
                  "tracks": [
                    {
                      "videoId": "rMbATaj7Il8",
                      "title": "Next Song",
                      "stream_url": "https://...",
                      "thumbnail": "https://..."
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Se requiere video_id o playlist_id"
          },
          "500": {
            "description": "Error interno"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stream/batch": {
      "get": {
        "tags": [
          "stream",
          "stream"
        ],
        "summary": "Get multiple stream URLs",
        "description": "Obtiene URLs de stream para múltiples videos. Usa MGET de Redis para optimizar cache como playlists.",
        "operationId": "get_batch_stream_urls_api_v1_stream_batch_get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lista de IDs separada por comas (máximo 50)",
              "title": "Ids"
            },
            "description": "Lista de IDs separada por comas (máximo 50)"
          },
          {
            "name": "bypass_cache",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Si true, ignora cache y obtiene URLs frescas de YouTube",
              "default": false,
              "title": "Bypass Cache"
            },
            "description": "Si true, ignora cache y obtiene URLs frescas de YouTube"
          }
        ],
        "responses": {
          "200": {
            "description": "Stream URLs obtenidas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Batch Stream Urls Api V1 Stream Batch Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Error de validación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "VALIDATION_ERROR",
                  "message": "Error de validación en 'video_id': ID de video inválido. Debe tener exactamente 11 caracteres.",
                  "details": {
                    "field": "video_id",
                    "reason": "invalid_length",
                    "expected_length": 11,
                    "actual_length": 5
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Error de autenticación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "AUTHENTICATION_ERROR",
                  "message": "Error de autenticación con YouTube Music. Verifica las credenciales.",
                  "details": {
                    "operation": "search",
                    "hint": "Check server logs for details"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Recurso no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "NOT_FOUND",
                  "message": "Video no encontrado.",
                  "details": {
                    "resource_type": "video",
                    "resource_id": "rMbATaj7Il8"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit excedido",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "RATE_LIMIT_ERROR",
                  "message": "Límite de peticiones excedido. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "retry_after": 300
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "INTERNAL_ERROR",
                  "message": "Ha ocurrido un error interno. Por favor intenta más tarde."
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway - Error de servicio externo",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "EXTERNAL_SERVICE_ERROR",
                  "message": "Error en YouTube Music durante get_stream. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "service": "YouTube Music"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Servicio no disponible",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Servicio temporalmente no disponible debido a sobrecarga.",
                  "details": {
                    "retry_after": 180,
                    "state": "OPEN"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stream/{video_id}": {
      "get": {
        "tags": [
          "stream",
          "stream"
        ],
        "summary": "Get audio stream URL",
        "description": "Obtiene la URL directa de stream de audio de una canción usando yt-dlp. Incluye caché inteligente y circuit breaker.",
        "operationId": "get_stream_url_api_v1_stream__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del video/canción",
              "title": "Video Id"
            },
            "description": "ID del video/canción"
          },
          {
            "name": "bypass_cache",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Si true, ignora la caché y genera una URL fresca",
              "default": false,
              "title": "Bypass Cache"
            },
            "description": "Si true, ignora la caché y genera una URL fresca"
          }
        ],
        "responses": {
          "200": {
            "description": "Stream URL obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Stream Url Api V1 Stream  Video Id  Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Error de validación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "VALIDATION_ERROR",
                  "message": "Error de validación en 'video_id': ID de video inválido. Debe tener exactamente 11 caracteres.",
                  "details": {
                    "field": "video_id",
                    "reason": "invalid_length",
                    "expected_length": 11,
                    "actual_length": 5
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Error de autenticación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "AUTHENTICATION_ERROR",
                  "message": "Error de autenticación con YouTube Music. Verifica las credenciales.",
                  "details": {
                    "operation": "search",
                    "hint": "Check server logs for details"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Recurso no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "NOT_FOUND",
                  "message": "Video no encontrado.",
                  "details": {
                    "resource_type": "video",
                    "resource_id": "rMbATaj7Il8"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit excedido",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "RATE_LIMIT_ERROR",
                  "message": "Límite de peticiones excedido. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "retry_after": 300
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "INTERNAL_ERROR",
                  "message": "Ha ocurrido un error interno. Por favor intenta más tarde."
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway - Error de servicio externo",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "EXTERNAL_SERVICE_ERROR",
                  "message": "Error en YouTube Music durante get_stream. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "service": "YouTube Music"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Servicio no disponible",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Servicio temporalmente no disponible debido a sobrecarga.",
                  "details": {
                    "retry_after": 180,
                    "state": "OPEN"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetailItem": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Field that caused the error",
            "examples": [
              "query",
              "video_id",
              "limit"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message",
            "examples": [
              "campo requerido",
              "valor inválido"
            ]
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type",
            "description": "Error type identifier",
            "examples": [
              "value_error.missing",
              "type_error.integer"
            ]
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetailItem",
        "description": "Single error detail item for validation errors."
      },
      "ErrorDetails": {
        "properties": {
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetailItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "List of individual errors (for validation errors)"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count",
            "description": "Number of errors"
          },
          "operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation",
            "description": "Operation that caused the error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "Field that caused the error"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Reason for the error"
          },
          "retry_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After",
            "description": "Seconds to wait before retrying"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type",
            "description": "Type of resource that was not found"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id",
            "description": "ID of resource that was not found"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "State (e.g., circuit breaker state)",
            "examples": [
              "OPEN",
              "CLOSED",
              "HALF_OPEN"
            ]
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint",
            "description": "Hint for resolving the error"
          }
        },
        "type": "object",
        "title": "ErrorDetails",
        "description": "Container for error details."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "boolean",
            "title": "Error",
            "description": "Always True for error responses",
            "default": true
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Machine-readable error code",
            "examples": [
              "VALIDATION_ERROR",
              "AUTHENTICATION_ERROR",
              "RATE_LIMIT_ERROR",
              "NOT_FOUND",
              "EXTERNAL_SERVICE_ERROR",
              "SERVICE_UNAVAILABLE",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message in Spanish",
            "examples": [
              "Error de validación en los parámetros de la petición",
              "Error de autenticación con YouTube Music",
              "Límite de peticiones excedido. Intenta más tarde.",
              "Recurso no encontrado",
              "Error en el servicio de YouTube Music",
              "Servicio temporalmente no disponible"
            ]
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional error details (structure varies by error type)"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema for all API errors.\n\nThis schema is used consistently across all endpoints to provide\na predictable error format for clients.\n\nAttributes:\n    error: Always True for error responses.\n    error_code: Machine-readable error code for programmatic handling.\n    message: Human-readable error message in Spanish.\n    details: Optional additional details about the error."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}